hss_app/app/requirements.txt
bromelia==0.3.1
psycopg2-binary==2.9.2
pycrypto==2.6.1
PyYAML==6.0
SQLAlchemy==1.4.27
